# server/src/openapi/api.yml
openapi: 3.0.1
info:
  title: API
  description: Documentation for the Cartable REST API routes.
  contact:
    email: frederic.olland@enseignementsup.gouv.fr
  license:
    name: MIT
    url: 'https://github.com/dataesr/cartable/LICENCE'
  version: 1.0.0
servers:
  - url: 'https://cartable.staging.dataesr.ovh/api'
    description: Staging server
  - url: 'https://cartable.dataesr.ovh/api'
    description: Production server
  - url: 'http://localhost:4000/api'
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  '/token':
    post:
      tags:
        - Authentication
      summary: |
        Allows to generate a new pair of access and refresh token by providing a valid refresh token.
        A new pair of token means a new refresh token whose expiration date is reseted with the configurated validity period.
        The current validity period of a refresh token is 20 days.
      security: []
      operationId: refreshToken
      requestBody:
        description: A valid refresh token
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  '/passwordless':
    post:
      tags:
        - Authentication
      summary: Recover a lost/forgotten/unset password.
      description: |
        Not passing X-Paysage-OTP returns 202 and triggers a new email with an OTP for the user to allow passord change.
      security: []
      operationId: passwordless
      parameters:
        - $ref: '#/components/parameters/XOtp'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                password:
                  type: string
                  pattern: '^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&:_])[A-Za-z\d@$!%*#?&:_]{8,}$'
                firstName:
                  type: string
                lastName:
                  type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        202:
          description: An email with an OTP has been sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'


  ### USERS ###
...
  '/folders':
    get:
      tags:
        - Folders
      summary: List folders
      operationId: listFolders
      security: []
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/Folder'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Folders
      summary: Create a new folder
      operationId: createFolder
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                access:
                  type: string
                  default: private
                  enum:
                    - private
                    - public
                isArchived:
                  type: boolean
                  default: false
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
...
    XOtp:
      name: X-Cartable-OTP
      in: header
      description: Use X-Cartable-OTP header to provide user's OTP
      schema:
        type: string
        pattern: '^[0-9]{6}$'
...
  responses:
    BadRequest:
      description: Illegal input for operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Permission denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication needed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
              path:
                type: string
              errorCode:
                type: string
...
    Folder:
      type: object
      additionalProperties: false
      properties:
        _id:
          type: string
          readOnly: true
          format: mongo-objectid
        name:
          type: string
        description:
          type: string
        isArchived:
          type: boolean
        access:
          type: string
          enum:
            - private
            - public
        isStarred:
          type: boolean
        viewerRole:
          type: string
          nullable: true
          enum:
            - null
            - admin
            - editor
            - viewer
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: string
          format: mongo-objectid
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        updatedBy:
          type: string
          format: mongo-objectid
          readOnly: true
...
    Tokens:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Authentication with a Bearer token.